requirements.txt
# ── Web Framework ──────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
jinja2==3.1.4
python-multipart==0.0.17      # required for Form() in FastAPI

# ── Data Validation ─────────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.7.0

# ── LLM ─────────────────────────────────────────────────────
google-genai==1.5.0            # google-generativeai new SDK (genai.Client)

# ── Database ─────────────────────────────────────────────────
# SQLite is built into Python — no package needed
# Optional AlloyDB (Cloud SQL PostgreSQL) driver:
# asyncpg==0.30.0
# cloud-sql-python-connector[asyncpg]==1.12.0

# ── Utilities ────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Dev / Testing (not needed in production) ─────────────────
# httpx==0.27.0                # for TestClient
# pytest==8.3.4
# pytest-asyncio==0.24.0
